+++ sentry-log/src/logger.rs
@@ -1,24 +1,27 @@
 use log::Record;
 use sentry_core::protocol::{Breadcrumb, Event};
 
+use bitflags::bitflags;
+
 #[cfg(feature = "logs")]
 use crate::converters::log_from_record;
 use crate::converters::{breadcrumb_from_record, event_from_record, exception_from_record};
 
-/// The action that Sentry should perform for a [`log::Metadata`].
-#[derive(Debug)]
-pub enum LogFilter {
-    /// Ignore the [`Record`].
-    Ignore,
-    /// Create a [`Breadcrumb`] from this [`Record`].
-    Breadcrumb,
-    /// Create a message [`Event`] from this [`Record`].
-    Event,
-    /// Create an exception [`Event`] from this [`Record`].
-    Exception,
-    /// Create a [`sentry_core::protocol::Log`] from this [`Record`].
-    #[cfg(feature = "logs")]
-    Log,
+bitflags! {
+    /// The action that Sentry should perform for a [`log::Metadata`].
+    #[derive(Debug, Clone, Copy)]
+    pub struct LogFilter: u32 {
+        /// Ignore the [`Record`].
+        const Ignore = 0b0000;
+        /// Create a [`Breadcrumb`] from this [`Record`].
+        const Breadcrumb = 0b0001;
+        /// Create a message [`Event`] from this [`Record`].
+        const Event = 0b0010;
+        /// Create an exception [`Event`] from this [`Record`].
+        const Exception = 0b0100;
+        /// Create a [`sentry_core::protocol::Log`] from this [`Record`].
+        const Log = 0b1000;
+    }
 }
 
 /// The type of Data Sentry should ingest for a [`log::Record`].
@@ -34,6 +37,28 @@ pub enum RecordMapping {
     /// Captures the [`sentry_core::protocol::Log`] to Sentry.
     #[cfg(feature = "logs")]
     Log(sentry_core::protocol::Log),
+    /// Captures multiple items to Sentry.
+    /// Nesting multiple `RecordMapping::Combined` inside each other will cause the inner mappings to be ignored.
+    Combined(CombinedRecordMapping),
+}
+
+/// A list of record mappings.
+#[derive(Debug)]
+pub struct CombinedRecordMapping(Vec<RecordMapping>);
+
+impl From<RecordMapping> for CombinedRecordMapping {
+    fn from(value: RecordMapping) -> Self {
+        match value {
+            RecordMapping::Combined(combined) => combined,
+            _ => CombinedRecordMapping(vec![value]),
+        }
+    }
+}
+
+impl From<Vec<RecordMapping>> for CombinedRecordMapping {
+    fn from(value: Vec<RecordMapping>) -> Self {
+        Self(value)
+    }
 }
 
 /// The default log filter.
@@ -132,30 +157,50 @@ impl<L: log::Log> SentryLogger<L> {
 
 impl<L: log::Log> log::Log for SentryLogger<L> {
     fn enabled(&self, metadata: &log::Metadata<'_>) -> bool {
-        self.dest.enabled(metadata) || !matches!((self.filter)(metadata), LogFilter::Ignore)
+        self.dest.enabled(metadata) || !((self.filter)(metadata).is_empty())
     }
 
     fn log(&self, record: &log::Record<'_>) {
-        let item: RecordMapping = match &self.mapper {
+        let items = match &self.mapper {
             Some(mapper) => mapper(record),
-            None => match (self.filter)(record.metadata()) {
-                LogFilter::Ignore => RecordMapping::Ignore,
-                LogFilter::Breadcrumb => RecordMapping::Breadcrumb(breadcrumb_from_record(record)),
-                LogFilter::Event => RecordMapping::Event(event_from_record(record)),
-                LogFilter::Exception => RecordMapping::Event(exception_from_record(record)),
+            None => {
+                let filter = (self.filter)(record.metadata());
+                let mut items = vec![];
+                if filter.contains(LogFilter::Breadcrumb) {
+                    items.push(RecordMapping::Breadcrumb(breadcrumb_from_record(record)));
+                }
+                if filter.contains(LogFilter::Event) {
+                    items.push(RecordMapping::Event(event_from_record(record)));
+                }
+                if filter.contains(LogFilter::Exception) {
+                    items.push(RecordMapping::Event(exception_from_record(record)));
+                }
                 #[cfg(feature = "logs")]
-                LogFilter::Log => RecordMapping::Log(log_from_record(record)),
-            },
-        };
-
-        match item {
-            RecordMapping::Ignore => {}
-            RecordMapping::Breadcrumb(b) => sentry_core::add_breadcrumb(b),
-            RecordMapping::Event(e) => {
-                sentry_core::capture_event(e);
+                if filter.contains(LogFilter::Log) {
+                    items.push(RecordMapping::Log(log_from_record(record)));
+                }
+                RecordMapping::Combined(CombinedRecordMapping(items))
             }
-            #[cfg(feature = "logs")]
-            RecordMapping::Log(log) => sentry_core::Hub::with_active(|hub| hub.capture_log(log)),
+        };
+        let items = CombinedRecordMapping::from(items);
+
+        for item in items.0 {
+            match item {
+                RecordMapping::Ignore => {}
+                RecordMapping::Breadcrumb(breadcrumb) => sentry_core::add_breadcrumb(breadcrumb),
+                RecordMapping::Event(event) => {
+                    sentry_core::capture_event(event);
+                }
+                #[cfg(feature = "logs")]
+                RecordMapping::Log(log) => {
+                    sentry_core::Hub::with_active(|hub| hub.capture_log(log))
+                }
+                RecordMapping::Combined(_) => {
+                    sentry_core::sentry_debug!(
+                        "[SentryLogger] found nested CombinedEventMapping, ignoring"
+                    )
+                }
+            }
         }
 
         self.dest.log(record)